Avoid SNAPSHOTs for release version. (#46)
According to https://docs.oracle.com/middleware/1212/core/MAVEN/maven_version.htm#MAVEN8855, "All versions with a qualifier are older than the same version without a qualifier (release version)." which means that specifying:
```
"chisel-iotesters" -> "[1.2.5,1.3.0["
```
as a version range, causes sbt-ivy to pull in 1.3-SNAPSHOT (now that we publish SNAPSHOTs) and eventually (due to dependencies), the SNAPSHOT versions of the rest of the BIG5. Use the fictitious "1.3.0-SNAPSHOT" version as the (exclusive) upper limit on the range.