Use release instead of snapshot
This changes build.sbt to fetch the newest 1.3.x and 0.1.y versions of
testers and testers2. This prevents problems related to
incompatibilities of mismatched dependencies of old snapshots and/or
the off-cadence snapshots of testers2.

Additionally, this removes chisel3 as a dependency since it will be
fetched by testers/testers2.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>